create the functionality for the singup page
